Busca de produtos por nome e categoria, e listagem geral

+++ src/main/java/com/marcus/cursomc/domain/Produto.java
@@ -1,7 +1,6 @@
 package com.marcus.cursomc.domain;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -28,7 +27,7 @@ public class Produto implements Serializable {
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	private Integer id;
 	private String nome;
-	private BigDecimal preco;
+	private Double preco;
 	
 	@JsonIgnore
 	@ManyToMany
@@ -46,7 +45,7 @@ public class Produto implements Serializable {
 		
 	}
 	
-	public Produto(Integer id, String nome, BigDecimal preco) {
+	public Produto(Integer id, String nome, Double preco) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -78,6 +77,7 @@ public class Produto implements Serializable {
 		return true;
 	}
 
+	
 	@JsonIgnore
 	public List<Pedido> getPedidos(){
 		List<Pedido> lista = new ArrayList<>();
@@ -103,11 +103,11 @@ public class Produto implements Serializable {
 		this.nome = nome;
 	}
 
-	public BigDecimal getPreco() {
+	public Double getPreco() {
 		return preco;
 	}
 
-	public void setPreco(BigDecimal preco) {
+	public void setPreco(Double preco) {
 		this.preco = preco;
 	}
